Position Sensor Added
Reads the data from a the linear position sensor, calculates the
displacement from the top in cm. Then it breaks the data up into the MSB
and LSB. MSB is data to the left of the decimal place, LSB is data to
the right of the decimal.

Additionally, the first optimizations for speed have been implemented in
the form of pointers in favor of Global variables.